requirements.txt
pytest~=6.2.5
pytest-mock
argparse~=1.4.0
mock~=4.0.3
PyYaml~=6.0
pycryptodome
aiohttp
pytest-aiohttp
pytest-asyncio